Commit (Final) Reparo e regras de negocios adicionadas, assim sendo o ultimo commit do projeto.

+++ src/services/carroService.ts
@@ -1,13 +1,13 @@
 import { Carro } from "../models/carro"
 import { CarroRepository } from "../repositories/carroRepository"
-// import { NotaFiscalRepository } from "../repositories/notaFiscalRepository"
-// import { EstoqueRepository } from "../repositories/estoqueRepository"
+import { NotaFiscalRepository } from "../repositories/notaFiscalRepository"
+import { EstoqueRepository } from "../repositories/estoqueRepository"
 import { ErroStatusRepository } from "../repositories/erroStatusRepository"
 
 export class CarroService {
     carroRepository: CarroRepository = CarroRepository.getInstance()
-    // notaRepository: NotaFiscalRepository = NotaFiscalRepository.getInstance()
-    // estoqueRepository: EstoqueRepository = EstoqueRepository.getInstance()
+    notaRepository: NotaFiscalRepository = NotaFiscalRepository.getInstance()
+    estoqueRepository: EstoqueRepository = EstoqueRepository.getInstance()
     erroStatus: ErroStatusRepository = ErroStatusRepository.getInstance()
 
     async insereCarro(carroBody: any): Promise<Carro> {
@@ -48,15 +48,13 @@ export class CarroService {
         return this.carroRepository.listaCarroID(Number(id))
     } 
 
-    // listaCarroDisponivel(): Carro[] | undefined {
-    //     const carroList = this.carroRepository.listaCarros()
-    //     const carroDisponiveis = carroList.filter(carro => this.estoqueRepository.listaCarroDisponivel(Number(carro.id_carro))?.quantidade != 0)
-    //     if(carroList.length === 0 || carroDisponiveis.length === 0){
-    //         this.erroStatus.insereErro(404)
-    //         throw new Error("Nenhum veiculo cadastrado")
-    //     }
-    //     return carroDisponiveis
-    // }
+    async listaCarroDisponivel(): Promise<any[] | undefined >{
+        if(await this.estoqueRepository.carrosDisponiveis() === undefined){
+            this.erroStatus.insereErro(404)
+            throw new Error("Nenhum veiculo cadastrado")
+        }
+        return this.estoqueRepository.carrosDisponiveis()
+    }
 
     async atualizaCarro(id: any, carroBody: any): Promise<Carro | undefined>{
         if(await this.carroRepository.listaCarroID(Number(id)) === undefined){
@@ -78,28 +76,21 @@ export class CarroService {
         return this.carroRepository.atualizaCarro(Number(id),carroBody)
     }
 
-    // deletaCarro(id: any): Carro[] | undefined {
-    //     if(this.carroRepository.listaCarroID(Number(id)) === undefined){
-    //         this.erroStatus.insereErro(404)
-    //         throw new Error("Carro com este ID, não está cadastrado no sistema.")
-    //     }
-    //     if(this.notaRepository.verificaNotaIDtabela(Number(id),"carro") != -1){
-    //         this.erroStatus.insereErro(422)
-    //         throw new Error("Carro não pode ser excluído por conta que tem nota emitida.")
-    //     }
-    //     if(this.estoqueRepository.listaEstoqueIDCarro(Number(id)) != undefined){
-    //         this.erroStatus.insereErro(422)
-    //         throw new Error("Carro não pode ser excluído por conta que tem estoque aberto.")
-    //     }
-    //     return this.carroRepository.deletaCarro(id)
-    // }
-
     async deleteCarro(id: any): Promise<Carro | undefined>{
         if(await this.carroRepository.listaCarroID(Number(id)) === undefined){
             this.erroStatus.insereErro(404)
             throw new Error("Carro com este ID, não está cadastrado no sistema.")
         }
+
+        if(await this.notaRepository.verificaCliente(Number(id)) != undefined){
+            this.erroStatus.insereErro(422)
+            throw new Error("Cliente não pode ser excluído por conta que tem nota emitida.")
+        }
         
+        if(this.estoqueRepository.listaEstoqueCarroID(Number(id)) != undefined){
+            this.erroStatus.insereErro(422)
+            throw new Error("Carro não pode ser excluído por conta que tem estoque aberto.")
+        }
         return this.carroRepository.deleteCarro(id)
     }
 }
