+++ src/services/clienteService.ts
@@ -1,11 +1,11 @@
 import { Cliente } from "../models/cliente"
 import { ClienteRepository } from "../repositories/clienteRepository"
-// import { NotaFiscalRepository } from "../repositories/notaFiscalRepository"
+import { NotaFiscalRepository } from "../repositories/notaFiscalRepository"
 import { NotaFiscal } from "../models/notaFiscal"
 import { ErroStatusRepository } from "../repositories/erroStatusRepository"
 export class ClienteService {
     clienteRepository: ClienteRepository = ClienteRepository.getInstance()
-    // notaRepository: NotaFiscalRepository = NotaFiscalRepository.getInstance()
+    notaRepository: NotaFiscalRepository = NotaFiscalRepository.getInstance()
     erroStatus: ErroStatusRepository = ErroStatusRepository.getInstance()
 
     async insereCliente(clienteBody: any): Promise<Cliente> {
@@ -45,13 +45,6 @@ export class ClienteService {
             throw new Error("Cliente com este ID, não existe no sistema.")
         }
 
-        // const indexId = this.clienteRepository.indexCliente(Number(id));
-        // const indexCPF = this.clienteRepository.cpfRepetido(clienteBody.cpf);
-        // if(indexCPF !== -1 && indexCPF !== indexId){
-        //     this.erroStatus.insereErro(409)
-        //     throw new Error("Não pode ter um vendedor com uma matrícula já cadastrada. Atualize com outra matrícula.")
-        // }
-
         return this.clienteRepository.atualizaCliente(Number(id),clienteBody)
     }
 
@@ -60,18 +53,18 @@ export class ClienteService {
             this.erroStatus.insereErro(404)
             throw new Error("Cliente com este ID, não está cadastrado no sistema.")
         }
-        // if(this.notaRepository.verificaNotaIDtabela(Number(id),"cliente") != -1){
-        //     this.erroStatus.insereErro(422)
-        //     throw new Error("Cliente não pode ser excluído por conta que tem nota emitida.")
-        // }
+        if(await this.notaRepository.verificaCliente(Number(id)) != undefined){
+            this.erroStatus.insereErro(422)
+            throw new Error("Cliente não pode ser excluído por conta que tem nota emitida.")
+        }
         return this.clienteRepository.deleteCliente(id)
     }
 
-//     listaNotasCliente(id: any): NotaFiscal[] | undefined {
-//         if(this.notaRepository.listaNotasporTabela(id,"cliente")!.length === 0){
-//             this.erroStatus.insereErro(404)
-//             throw new Error("Cliente não possuí notas emitidas no sistema.")
-//         }
-//         return this.notaRepository.listaNotasporTabela(id,"cliente")
-//     }
+    async listaNotasCliente(id: any): Promise<NotaFiscal[] | undefined >{
+        if(await this.notaRepository.listaNotasCliente(id) === undefined){
+            this.erroStatus.insereErro(404)
+            throw new Error("Cliente não possuí notas emitidas no sistema.")
+        }
+        return this.notaRepository.listaNotasCliente(id)
+    }
 }
